Core/DataStores: Updated structures to 9.2.5
Port From (https://github.com/TrinityCore/TrinityCore/commit/6c8ef045a52723618b91c565ae6c3efc711e7581)

+++ Source/Game/DataStorage/Structs/A_Records.cs
@@ -60,7 +60,7 @@ namespace Game.DataStorage
         public string ContinueDescription;
         public byte Type;
         public uint PlayerConditionID;
-        public byte Flags;
+        public int Flags;
         public byte ButtonActionType;
         public int TextureFileDataID;
         public ushort LfgDungeonID;
